Update createBidQuery.js, readAuctionEvents.js, and readAuctionProducts.js
Editing scripts slightly for clarity. I've gotten a bid to be placed with the current setup in createBidQuery, but because of the way the script is currently set up, it doesn't properly report the success to the front end (I checked the log and results in the DB). Will step away and try the admin reports this evening.

Feel free to to let me know if there's an elegant fix, esp to createBid.

Also, I am not particularly experienced with the components + Formik, and I don't have a bid placement form to use.